refactor(agent): extract PiAdapter's bridge script to a real file
BRIDGE_SCRIPT was a ~540-line Node.js program written as a JS
template literal inside pi.ts — no type checking, no lint, no
tooling support of its own. Moved verbatim into pi-bridge.js (a
real .js file, now actually linted/formatted).

Bun's native text import attribute (`with { type: "text" }`) was
tried first since it's the cleanest option and typechecks fine, but
this package's actual publish build goes through tsdown/rolldown,
which understands neither that nor the `?raw` convention - both
failed a real `bun run build`. Landed on reading the file at
runtime relative to its own module location, with tsdown's `copy`
config placing it next to dist/index.mjs so the same relative path
resolves in both dev and the published package. Verified against
the built dist/, not just source.

Content verified byte-identical to the original template literal
via two independent extraction methods (real JS evaluation and a
manual unescape) before switching pi.ts over.

Part of the codebase restructure plan (plans/codebase-restructure.md,
Phase 4).
