+++ packages/agent/tsdown.config.ts
@@ -8,4 +8,9 @@ export default defineConfig({
   platform: "node",
   clean: true,
   deps: { neverBundle: ["@drej/core", "drej", "@drej/sqlite"] },
+  // pi-bridge.js is read at runtime relative to this module's own location (see
+  // adapters/pi.ts) rather than bundled as a string — copy it alongside index.mjs
+  // so that resolution works identically in dev (src/adapters/) and in the
+  // published package (dist/).
+  copy: ["src/adapters/pi-bridge.js"],
 });
